# public/openapi.yaml
openapi: 3.0.3
info:
  title: HPKV Cloud API
  description: API endpoints for interacting with an HPKV store, mirroring hpkvfs logic.
  version: 0.1.0
servers:
  - url: /api # Relative path to the API routes within the Next.js app
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: hpkv-api-key
      description: Your HPKV API Key.
    ApiUrlAuth:
      type: apiKey
      in: header
      name: hpkv-api-url
      description: The base URL of your HPKV API endpoint (e.g., https://api-eu-1.hpkv.io).
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Description of the error.
security:
  - ApiKeyAuth: []
  - ApiUrlAuth: []
paths:
...
  /read:
    get:
      summary: Read file content
      description: Reads a specified range of bytes from a file.
      parameters:
        - name: path
          in: query
          required: true
          description: The absolute path of the file to read.
          schema:
            type: string
        - name: offset
          in: query
          required: true
          description: The byte offset at which to start reading.
          schema:
            type: integer
            format: int64
            minimum: 0
        - name: size
          in: query
          required: true
          description: The maximum number of bytes to read.
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '200':
          description: Successfully read file content.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: Bad Request (e.g., missing parameters, invalid offset/size, path is directory).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found (file does not exist).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /mkdir:
    post:
      summary: Create directory
      description: Creates a new directory at the specified path.
      parameters:
        - name: path
          in: query
          required: true
          description: The absolute path of the directory to create. Should not end with a slash.
          schema:
            type: string
      responses:
        '200':
          description: Directory already exists (idempotent success).
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '201':
          description: Successfully created the directory.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          description: Bad Request (e.g., missing path, invalid path format).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict (path exists but is not a directory).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'